+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
-    <link href="/src/assets/style/_custom.css" rel="stylesheet" />
+    <link href="/src/assets/style/_custom-properties.css" rel="stylesheet" />
     <link href="/src/assets/style/index.css" rel="stylesheet" />
     <link href="/src/assets/style/application.css" rel="stylesheet" />
     <link href="/src/assets/style/new-transaction-form.css" rel="stylesheet" />
@@ -22,7 +22,7 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
+    <div id="application-root"></div>
 
     <script src="/src/index.tsx" type="module"></script>
   </body>
