+++ src/assets/style/index.css
@@ -15,109 +15,3 @@ body {
   background: var(--background);
   font-size: var(--font-size);
 }
-#root {
-  display: contents;
-}
-
-header {
-  width: 100%;
-  height: var(--header-height);
-  display: flex;
-
-  justify-content: space-between;
-  align-items: center;
-  justify-items: center;
-  width: 100%;
-  background-color: var(--midground);
-  padding: 8px;
-  font-size: 24rem;
-}
-
-.not-showing {
-  background-color: var(--midground);
-}
-.bezos-colored {
-  color: red;
-}
-
-.transaction-list tr:first-child {
-  position: sticky;
-  background-color: var(--background);
-
-  top: calc(-1 * var(--padding));
-}
-
-.transaction-list tr.bezos {
-  border-radius: 8px;
-  padding: var(--padding-minor);
-  box-shadow: var(--midground) 0 2px 5px 0;
-}
-.transaction-list tr.not-showing.bezos {
-  color: darkred;
-}
-
-.transaction-list tr td {
-  text-align: center;
-}
-
-.transaction-list button.hide::before {
-  content: "Hide";
-}
-
-.transaction-list tr.not-showing button.hide::before {
-  content: "Show";
-}
-
-.transaction-list button.bezoslist::before {
-  content: "+";
-}
-
-.transaction-list tr.bezos button.bezoslist::before {
-  content: "-";
-}
-
-.transaction-list button.hide::before {
-  content: "Hide";
-}
-
-.transaction-list tr.not-showing button.hide::before {
-  content: "Show";
-}
-
-.transaction-list button.bezoslist::before {
-  content: "+";
-}
-
-.transaction-list tr.bezos button.bezoslist::before {
-  content: "-";
-}
-
-/*
-// ////////////////
-Totals Counter
-// ////////////////
-*/
-
-.horizontal-bar {
-  width: 80px;
-  height: 16rem;
-  background-color: var(--midground);
-  border: 1px solid var(--midground);
-  display: inline-block;
-}
-.horizontal-bar .bar {
-  display: block;
-  background-color: red;
-}
-
-.color-scheme-button::before {
-  content: "◠";
-}
-
-.dark .color-scheme-button::before {
-  content: "☾";
-}
-
-.light .color-scheme-button::before {
-  content: "☼";
-}
