+++ src/assets/style/transaction-list.css
@@ -41,6 +41,58 @@ Transaction List
   box-shadow: var(--midground) 0 2px 5px 0;
 }
 
+.transaction-list tr:first-child {
+  position: sticky;
+  background-color: var(--background);
+
+  top: calc(-1 * var(--padding));
+}
+
+.transaction-list tr.bezos {
+  border-radius: 8px;
+  padding: var(--padding-minor);
+  box-shadow: var(--midground) 0 2px 5px 0;
+}
+.transaction-list tr.not-showing.bezos {
+  color: darkred;
+}
+
+.transaction-list tr td {
+  text-align: center;
+}
+
+.transaction-list button.hide::before {
+  content: "Hide";
+}
+
+.transaction-list tr.not-showing button.hide::before {
+  content: "Show";
+}
+
+.transaction-list button.bezoslist::before {
+  content: "+";
+}
+
+.transaction-list tr.bezos button.bezoslist::before {
+  content: "-";
+}
+
+.transaction-list button.hide::before {
+  content: "Hide";
+}
+
+.transaction-list tr.not-showing button.hide::before {
+  content: "Show";
+}
+
+.transaction-list button.bezoslist::before {
+  content: "+";
+}
+
+.transaction-list tr.bezos button.bezoslist::before {
+  content: "-";
+}
+
 .transaction-list td.category {
   text-overflow: ellipsis;
   overflow: hidden;
