+++ src/index.tsx
@@ -2,6 +2,9 @@
 import { render } from "solid-js/web";
 
 // import "./index.css";
-import App from "./App";
+import Application from "./Application";
 
-render(() => <App />, document.getElementById("root") as HTMLElement);
+render(
+  () => <Application />,
+  document.getElementById("application-root") as HTMLElement
+);
